+++ src/hkimjp/wil2/todays.clj
@@ -7,8 +7,9 @@
    [ring.util.anti-forgery :refer [anti-forgery-field]]
    [ring.util.response :as resp]
    [taoensso.telemere :as t]
+   [hkimjp.carmine :as c]
    [hkimjp.datascript :as ds]
-   [hkimjp.wil2.util :refer [user today abbrev]]
+   [hkimjp.wil2.util :refer [user today]]
    [hkimjp.wil2.view :refer [page]]))
 
 (def uploaded? '[:find ?e
@@ -25,7 +26,8 @@
                       [?e :login ?login]
                       [?e :date ?today]])
 
-(defn upload [request]
+(defn upload
+  [request]
   (t/log! :debug "upload")
   (let [uploaded (ds/qq todays-uploads (today))]
     (page
@@ -58,7 +60,8 @@
                   :date (today)
                   :updated (jt/local-date-time)})
         (page [:div "upload success."]))
-      (page [:div "did not select a file to upload."]))))
+      (page [:div.text-red-600 "error"
+             [:p "did not select a file to upload."]]))))
 
 (defn- pt [s]
   (condp = (last (str/split s #"/"))
@@ -66,22 +69,29 @@
     "soso" 1
     "bad" -1))
 
-(defn point [{params :params :as request}]
-  (t/log! :info "point")
-  (t/log! :info (str "params " params))
-  (t/log! :info (str "uri: " (:uri request)))
-  (ds/put! {:wil2 "point"
-            :login (user request)
-            :to/id (parse-long (:eid params))
-            :pt (pt (:uri request))
-            :updated (jt/local-date-time)})
-  (resp/response "<p>received</p>"))
+;; ここで redis にメモる。
+(defn point! [{params :params :as request}]
+  (let [user (user request)
+        id (parse-long (:eid params))
+        pt (pt (:uri request))]
+    (t/log! :info (str "point! " user " to " id " pt " pt))
+    (ds/put! {:wil2 "point"
+              :login user
+              :to/id id
+              :pt pt
+              :updated (jt/local-date-time)})
+    (resp/response "<p>received</p>")))
 
-(defn md [{{:keys [eid]} :path-params :as request}]
+(defn- button [key sym]
+  [:button {:hx-post (str "/wil2/point/" key)
+            :hx-target "#wil"}
+   [:span.hover:text-2xl sym]])
+
+(defn md
+  "get /wil2/md/:eid"
+  [{{:keys [eid]} :path-params :as request}]
   (let [md (:md (ds/pl (parse-long eid)))
-        markdown (-> md
-                     md/parse
-                     md/->hiccup)]
+        markdown (-> md md/parse md/->hiccup)]
     (t/log! :info (str "md " (user request) " " eid))
     (resp/response
      (str (h/html
@@ -91,15 +101,9 @@
             markdown
             [:div.flex.gap-x-4
              [:span.py-2.font-bold "評価: "]
-             [:button {:hx-post "/wil2/point/good"
-                       :hx-target "#wil"}
-              [:span.hover:text-2xl "⬆️"]]
-             [:button {:hx-post "/wil2/point/soso"
-                       :hx-target "#wil"}
-              [:span.hover:text-2xl "➡️"]]
-             [:button {:hx-post "/wil2/point/bad"
-                       :hx-target "#wil"}
-              [:span.hover:text-2xl "⬇️"]]]])))))
+             (button "good" "⬆️")
+             (button "soso" "➡️")
+             (button "bad"  "⬇️")]])))))
 
 (defn- link [[eid login]]
   [:span.px-2.hover:underline
@@ -107,7 +111,8 @@
     :hx-target "#wil"}
    login])
 
-(defn todays [request]
+(defn todays
+  [request]
   (t/log! :debug "todays")
   (let [uploads (ds/qq todays-uploads (today))]
     (page
@@ -116,7 +121,7 @@
       [:p "他のユーザの WIL を読んで評価する。"]
       [:div.font-bold "uploaded"]
       (into [:div.mx-2] (mapv link uploads))
-      [:div#wil.py-2 [:span.font-bold "評価:"] "　⬆️　➡️　⬇️"]])))
+      [:div#wil.py-2 [:span.font-bold "評価: "] " ⬆️ ➡️ ⬇️"]])))
 
 (defn switch [request]
   (t/log! :debug "switch")